+++ projects/igniteui-angular/src/lib/calendar/calendar.component.ts
@@ -134,31 +134,19 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
     }
 
     /**
-     * Gets the `locale` of the calendar.
+     * Sets/gets the `locale` of the calendar.
+     * Expects a valid BCP 47 language tag.
      * Default value is `"en"`.
+     * ```html
+     * <igx-calendar [locale] = "de"></igx-calendar>
+     * ```
      * ```typescript
      * let locale =  this.calendar.locale;
      * ```
      * @memberof IgxCalendarComponent
      */
     @Input()
-    public get locale(): string {
-        return this._locale;
-    }
-
-    /**
-     * Sets the `locale` of the calendar.
-     * Expects a valid BCP 47 language tag.
-     * Default value is `"en"`.
-     * ```html
-     * <igx-calendar [locale] = "de"></igx-calendar>
-     * ```
-     * @memberof IgxCalendarComponent
-     */
-    public set locale(value: string) {
-        this._locale = value;
-        this.initFormatters();
-    }
+    public locale = 'en';
 
     /**
      *
@@ -219,7 +207,7 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
      * @memberof IgxCalendarComponent
      */
     public set viewDate(value: Date) {
-        this._viewDate = new Date(value);
+        this._viewDate = this.getDateOnly(value);
     }
 
     /**
@@ -271,7 +259,6 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
      */
     public set formatOptions(formatOptions: object) {
         this._formatOptions = Object.assign(this._formatOptions, formatOptions);
-        this.initFormatters();
     }
 
     /**
@@ -601,10 +588,6 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
      */
     private _monthAction = '';
     /**
-    *@hidden
-    */
-    private _locale = 'en';
-    /**
      *@hidden
      */
     private _formatOptions = {
@@ -628,26 +611,6 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
     /**
      *@hidden
      */
-    private formatterMonth;
-    /**
-     *@hidden
-     */
-    private formatterDay;
-    /**
-     *@hidden
-     */
-    private formatterYear;
-    /**
-     *@hidden
-     */
-    private formatterMonthday;
-    /**
-     *@hidden
-     */
-    private formatterWeekday;
-    /**
-     *@hidden
-     */
     private _specialDates: DateRangeDescriptor[] = null;
     /**
      * @hidden
@@ -664,20 +627,6 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
 
         this.calendarModel.firstWeekDay = this.weekStart;
         this._viewDate = this._viewDate ? this._viewDate : today;
-        this.initFormatters();
-    }
-
-    /**
-     * Resets the formatters when locale or formatOptions are changed
-     *
-     * @hidden
-     */
-    private initFormatters() {
-        this.formatterMonth = new Intl.DateTimeFormat(this._locale, { month: this._formatOptions.month });
-        this.formatterDay = new Intl.DateTimeFormat(this._locale, { day: this._formatOptions.day });
-        this.formatterYear = new Intl.DateTimeFormat(this._locale, { year: this._formatOptions.year });
-        this.formatterMonthday = new Intl.DateTimeFormat(this._locale, { month: this._formatOptions.month, day: this._formatOptions.day });
-        this.formatterWeekday = new Intl.DateTimeFormat(this._locale, { weekday: this._formatOptions.weekday });
     }
 
     /**
@@ -709,7 +658,7 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
      */
     public formattedMonth(value: Date): string {
         if (this._formatViews.month) {
-            return this.formatterMonth.format(value);
+            return value.toLocaleString(this.locale, { month: this._formatOptions.month });
         }
         return `${value.getMonth()}`;
     }
@@ -722,7 +671,7 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
      */
     public formattedDate(value: Date): string {
         if (this._formatViews.day) {
-            return this.formatterDay.format(value);
+            return value.toLocaleString(this.locale, { day: this._formatOptions.day });
         }
         return `${value.getDate()}`;
     }
@@ -735,7 +684,7 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
      */
     public formattedYear(value: Date): string {
         if (this._formatViews.year) {
-            return this.formatterYear.format(value);
+            return value.toLocaleString(this.locale, { year: this._formatOptions.year });
         }
         return `${value.getFullYear()}`;
     }
@@ -792,8 +741,9 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
         const date = this.headerDate;
 
         return {
-            monthday: this.formatterMonthday.format(date),
-            weekday: this.formatterWeekday.format(date),
+            monthday: date.toLocaleString(
+                this.locale, { month: this._formatOptions.month, day: this._formatOptions.day }),
+            weekday: date.toLocaleString(this.locale, { weekday: this._formatOptions.weekday })
         };
     }
 
@@ -814,12 +764,11 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
     }
 
     public animationDone(event, isLast: boolean) {
-        if (isLast) {
-            const date = this.dates.find((d) => d.selected);
-            if (date) {
-                setTimeout(() => date.nativeElement.focus(),
-                    parseInt(slideInRight.options.params.duration, 10));
-            }
+        const date = this.dates.find((d) => d.selected);
+
+        if (date && isLast) {
+            setTimeout(() => date.nativeElement.focus(),
+                parseInt(slideInRight.options.params.duration, 10));
         }
     }
 
@@ -905,7 +854,7 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
         const dayNames = [];
         const rv = this.calendarModel.monthdatescalendar(this.viewDate.getFullYear(), this.viewDate.getMonth())[0];
         for (const day of rv) {
-            dayNames.push(this.formatterWeekday.format(day.date));
+            dayNames.push(day.date.toLocaleString(this.locale, { weekday: this._formatOptions.weekday }));
         }
 
         return dayNames;
@@ -922,7 +871,7 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
      * @hidden
      */
     public changeYear(event: Date) {
-        this._viewDate = new Date(event.getFullYear(), this._viewDate.getMonth(), 1, 0, 0, 0);
+        this._viewDate = new Date(event.getFullYear(), this._viewDate.getMonth(), 1);
         this._activeView = CalendarView.DEFAULT;
     }
 
@@ -930,7 +879,7 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
      * @hidden
      */
     public changeMonth(event: Date) {
-        this._viewDate = new Date(this._viewDate.getFullYear(), event.getMonth(), 1, 0, 0, 0);
+        this._viewDate = new Date(this._viewDate.getFullYear(), event.getMonth(), 1);
         this._activeView = CalendarView.DEFAULT;
     }
 
@@ -1130,7 +1079,7 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
      * @hidden
      */
     private selectSingle(value: Date) {
-        this.selectedDates = value;
+        this.selectedDates = this.getDateOnly(value);
         this._onChangeCallback(this.selectedDates);
     }
 
@@ -1140,13 +1089,14 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
      */
     private selectMultiple(value: Date | Date[]) {
         if (Array.isArray(value)) {
-            this.selectedDates = this.selectedDates.concat(value);
+            this.selectedDates = this.selectedDates.concat(value.map(v => this.getDateOnly(v)));
         } else {
-            if (this.selectedDates.every((date: Date) => date.toDateString() !== value.toDateString())) {
-                this.selectedDates.push(value);
+            const valueDateOnly = this.getDateOnly(value);
+            if (this.selectedDates.every((date: Date) => date.getTime() !== valueDateOnly.getTime())) {
+                this.selectedDates.push(valueDateOnly);
             } else {
                 this.selectedDates = this.selectedDates.filter(
-                    (date: Date) => date.toDateString() !== value.toDateString()
+                    (date: Date) => date.getTime() !== valueDateOnly.getTime()
                 );
             }
         }
@@ -1163,17 +1113,18 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
         if (Array.isArray(value)) {
             this.rangeStarted = false;
             value.sort((a: Date, b: Date) => a.valueOf() - b.valueOf());
-            start = value.shift();
-            end = value.pop();
+            start =  this.getDateOnly(value.shift());
+            end =  this.getDateOnly(value.pop());
             this.selectedDates = [start, ...this.generateDateRange(start, end)];
         } else {
+            value = this.getDateOnly(value);
             if (!this.rangeStarted) {
                 this.rangeStarted = true;
                 this.selectedDates = [value];
             } else {
                 this.rangeStarted = false;
 
-                if (this.selectedDates[0].toDateString() === value.toDateString()) {
+                if (this.selectedDates[0].getTime() === value.getTime()) {
                     this.selectedDates = [];
                     this._onChangeCallback(this.selectedDates);
                     return;
@@ -1247,8 +1198,8 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
         }
 
         value.sort((a: Date, b: Date) => a.valueOf() - b.valueOf());
-        const start = value.shift();
-        const end = value.pop();
+        const start =  this.getDateOnly(value.shift());
+        const end =  this.getDateOnly(value.pop());
 
         const deselectRange = [start, ...this.generateDateRange(start, end)];
         this.deselectMultiple(deselectRange);
@@ -1275,12 +1226,12 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
      *@hidden
      */
     private isDateInRanges(date: Date, ranges: DateRangeDescriptor[]): boolean {
-        date = new Date(date.getFullYear(), date.getMonth(), date.getDate());
+        date = this.getDateOnly(date);
         const dateInMs = date.getTime();
 
         for (const descriptor of ranges) {
             const dRanges = descriptor.dateRange ? descriptor.dateRange.map(
-                r => new Date(r.getFullYear(), r.getMonth(), r.getDate())) : undefined;
+                r => this.getDateOnly(r)) : undefined;
             switch (descriptor.type) {
                 case (DateRangeType.After):
                     if (dateInMs > dRanges[0].getTime()) {
@@ -1353,8 +1304,9 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
      */
     private generateDateRange(start: Date, end: Date): Date[] {
         const result = [];
-
-        while (start.toDateString() !== end.toDateString()) {
+        start = this.getDateOnly(start);
+        end = this.getDateOnly(end);
+        while (start.getTime() !== end.getTime()) {
             start = this.calendarModel.timedelta(start, 'day', 1);
             result.push(start);
         }
@@ -1377,7 +1329,14 @@ export class IgxCalendarComponent implements OnInit, ControlValueAccessor {
      *@hidden
      */
     private getDateOnlyInMs(date: Date) {
-        return new Date(date.getFullYear(), date.getMonth(), date.getDate()).getTime();
+        return this.getDateOnly(date).getTime();
+    }
+
+     /**
+     *@hidden
+     */
+    private getDateOnly(date: Date) {
+        return new Date(date.getFullYear(), date.getMonth(), date.getDate());
     }
     /**
      *@hidden