+++ projects/igniteui-angular/src/lib/navigation-drawer/navigation-drawer.component.spec.ts
@@ -6,8 +6,9 @@ import {
   tick
 } from '@angular/core/testing';
 
-import { Component, ViewChild } from '@angular/core';
+import { Component, DebugElement, ViewChild } from '@angular/core';
 import { By } from '@angular/platform-browser';
+import { Observable } from 'rxjs';
 import * as Infragistics from '../../public_api';
 import { wait } from '../test-utils/ui-interactions.spec';
 
@@ -29,7 +30,6 @@ describe('Navigation Drawer', () => {
                 ],
                 imports: [Infragistics.IgxNavigationDrawerModule]
             });
-
             // Using Window through DI causes AOT error (https://github.com/angular/angular/issues/15640)
             // so for tests just force override the the `getWindowWidth`
             this.widthSpyOverride = spyOn(Infragistics.IgxNavigationDrawerComponent.prototype as any, 'getWindowWidth')
@@ -107,7 +107,6 @@ describe('Navigation Drawer', () => {
             set: {
                 template
             }});
-
             // compile after overrides, not in before each: https://github.com/angular/angular/issues/10712
             TestBed.compileComponents().then(() => {
                 const fixture = TestBed.createComponent(TestComponentDIComponent);
@@ -157,7 +156,6 @@ describe('Navigation Drawer', () => {
         it('async API calls should emit events', async(() => {
             let fixture: ComponentFixture<TestComponentDIComponent>;
             let drawer;
-
             // compile after overrides, not in before each: https://github.com/angular/angular/issues/10712
             TestBed.compileComponents().then(() => {
                 fixture = TestBed.createComponent(TestComponentDIComponent);
@@ -197,7 +195,6 @@ describe('Navigation Drawer', () => {
             set: {
                 template
             }});
-
             // compile after overrides, not in before each: https://github.com/angular/angular/issues/10712
             TestBed.compileComponents().then(() => {
                 const fixture = TestBed.createComponent(TestComponentDIComponent);
@@ -212,7 +209,6 @@ describe('Navigation Drawer', () => {
         }));
 
         it('should update with dynamic min template', async(() => {
-
             // immediate requestAnimationFrame for testing
             spyOn(window, 'requestAnimationFrame').and.callFake((callback) => callback());
             const template = `<igx-nav-drawer>
@@ -225,7 +221,6 @@ describe('Navigation Drawer', () => {
             }});
             let fixture;
             let asideElem;
-
             // compile after overrides, not in before each: https://github.com/angular/angular/issues/10712
             TestBed.compileComponents().then(() => {
                 fixture = TestBed.createComponent(TestComponentMini);
@@ -254,7 +249,6 @@ describe('Navigation Drawer', () => {
             set: {
                 template
             }});
-
             // compile after overrides, not in before each: https://github.com/angular/angular/issues/10712
             TestBed.compileComponents().then(() => {
                 const fixture = TestBed.createComponent(TestComponentPin);
@@ -376,7 +370,6 @@ describe('Navigation Drawer', () => {
                 expect(fixture.componentInstance.viewChild.isOpen).toEqual(false);
 
                 const listener = navDrawer.renderer.listen(document.body, 'panmove', () => {
-
                     // mid gesture
                     expect(navDrawer.drawer.classList).toContain('panning');
                     expect(navDrawer.drawer.style.transform)
@@ -388,17 +381,14 @@ describe('Navigation Drawer', () => {
             })
             .then(() => {
                 expect(navDrawer.isOpen).toEqual(false, 'should ignore too short pan');
-
                 // valid pan
                 return pan(document.body, 10, 10, 100, 200, 0);
             }).then(() => {
                 expect(navDrawer.isOpen).toEqual(true, 'should open on valid pan');
-
                 // not enough distance, closing
                 return pan(document.body, 200, 10, 100, -20, 0);
             }).then(() => {
                 expect(navDrawer.isOpen).toEqual(true, 'should remain open on too short pan');
-
                 // close
                 return pan(document.body, 250, 10, 100, -200, 0);
             }).then(() => {
@@ -419,11 +409,11 @@ describe('Navigation Drawer', () => {
             set: {
                 template
             }});
-
             // compile after overrides, not in before each: https://github.com/angular/angular/issues/10712
             TestBed.compileComponents().then(() => {
                 fixture = TestBed.createComponent(TestComponentDIComponent);
                 fixture.detectChanges();
+                const drawer: Infragistics.IgxNavigationDrawerComponent = fixture.componentInstance.viewChild;
 
                 fixture.componentInstance.drawerMiniWidth = 60;
                 fixture.detectChanges();
@@ -449,7 +439,6 @@ describe('Navigation Drawer', () => {
             set: {
                 template
             }});
-
             // compile after overrides, not in before each: https://github.com/angular/angular/issues/10712
             TestBed.compileComponents().then(() => {
                 fixture = TestBed.createComponent(TestComponentDIComponent);
@@ -498,11 +487,9 @@ describe('Navigation Drawer', () => {
             // compile after overrides, not in before each: https://github.com/angular/angular/issues/10712
             await TestBed.compileComponents();
             fixture = TestBed.createComponent(TestComponentPin);
-
             // watch for initial pin with 2-way bind expression changed errors
             expect(() => fixture.detectChanges()).not.toThrow();
             await fixture.whenStable();
-
             // defaults:
             expect(fixture.componentInstance.viewChild.pin).toBe(false, 'Should be initially unpinned');
             expect(fixture.componentInstance.pin).toBe(false, 'Parent component pin should update initially');
@@ -511,7 +498,6 @@ describe('Navigation Drawer', () => {
             fixture.componentInstance.pin = true;
             fixture.detectChanges();
             window.dispatchEvent(new Event('resize'));
-
             // wait for debounce
             await wait(200);
             expect(fixture.componentInstance.viewChild.pin).toBe(false, `Shouldn't change state on resize if window width is the same`);
@@ -521,7 +507,6 @@ describe('Navigation Drawer', () => {
 
             this.widthSpyOverride.and.returnValue(fixture.componentInstance.pinThreshold);
             window.dispatchEvent(new Event('resize'));
-
             // wait for debounce
             await wait(200);
             expect(fixture.componentInstance.viewChild.pin).toBe(true, 'Should pin on window resize over threshold');
@@ -529,7 +514,6 @@ describe('Navigation Drawer', () => {
 
             this.widthSpyOverride.and.returnValue(768);
             window.dispatchEvent(new Event('resize'));
-
             // wait for debounce
             await wait(200);
             expect(fixture.componentInstance.viewChild.pin).toBe(false, 'Should un-pin on window resize below threshold');
@@ -544,7 +528,6 @@ describe('Navigation Drawer', () => {
 
         it('should get correct window width', (done) => {
             const originalWidth = window.innerWidth;
-
             // re-enable `getWindowWidth`
             const widthSpy = (this.widthSpyOverride as jasmine.Spy).and.callThrough();
             expect(widthSpy.call(null)).toEqual(originalWidth);
@@ -563,7 +546,6 @@ describe('Navigation Drawer', () => {
             };
 
             return new Promise((resolve, reject) => {
-
                 // force touch (https://github.com/hammerjs/hammer.js/issues/1065)
                 Simulator.setType('touch');
                 Simulator.gestures.swipe(element, swipeOptions, () => {
@@ -581,7 +563,6 @@ describe('Navigation Drawer', () => {
             };
 
             return new Promise((resolve, reject) => {
-
                 // force touch (https://github.com/hammerjs/hammer.js/issues/1065)
                 Simulator.setType('touch');
                 Simulator.gestures.pan(element, swipeOptions, () => {